fix: use Expr for any A_Expr/FuncCall-accepting field
Since the Expr type consists of a ton of variants, there won‘t exist enough test cases for inference to get us all the way there. If a field is known to possibly be (or include) an A_Expr or FuncCall node, we can be 95% sure it supports any Expr node.

Closes #13